+++ crates/env/src/call/call_builder/call_v1.rs
@@ -29,7 +29,10 @@ use crate::{
     Gas,
 };
 use num_traits::Zero;
+#[cfg(not(feature = "revive"))]
 use pallet_contracts_uapi::CallFlags;
+#[cfg(feature = "revive")]
+use pallet_revive_uapi::CallFlags;
 
 /// The legacy call type for cross-contract calls. Performs a cross-contract call to
 /// `callee` with gas limit `gas_limit`, transferring `transferred_value` of currency.