+++ crates/env/src/error.rs
@@ -16,7 +16,10 @@ use derive_more::From;
 
 #[cfg(any(feature = "std", test, doc))]
 use crate::engine::off_chain::OffChainError;
+#[cfg(not(feature = "revive"))]
 use pallet_contracts_uapi::ReturnErrorCode;
+#[cfg(feature = "revive")]
+use pallet_revive_uapi::ReturnErrorCode;
 
 /// Errors that can be encountered upon environmental interaction.
 #[derive(Debug, From, PartialEq, Eq)]