+++ crates/env/src/lib.rs
@@ -102,8 +102,16 @@ mod tests;
 #[doc(inline)]
 pub use self::engine::off_chain::test_api as test;
 
+#[cfg(not(feature = "revive"))]
 #[doc(inline)]
 pub use pallet_contracts_uapi::{
+    CallFlags,
+    ReturnErrorCode,
+    ReturnFlags,
+};
+#[cfg(feature = "revive")]
+#[doc(inline)]
+pub use pallet_revive_uapi::{
     CallFlags,
     ReturnErrorCode,
     ReturnFlags,